server/src/main/java/com/tpi/server/application/usecases/user/GetUserDataUseCase.java
package com.tpi.server.application.usecases.user;

import com.tpi.server.domain.models.User;
import com.tpi.server.infrastructure.repositories.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;

@Component
@RequiredArgsConstructor
public class GetUserDataUseCase {

    private final UserRepository userRepository;

    public User execute(Integer userId) {
        return userRepository.findById(userId)
                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
    }
}
